Move service messages to contracts
