+++ source/Calamari.Contracts/ArgoCD/ArgoCDCustomPropertiesDto.cs
